backend/lib/supabase_client.py
"""
Supabase client for FastAPI backend.
Uses the supabase-py library with service role key for admin operations.
"""
import os
from supabase import create_client, Client
from dotenv import load_dotenv

load_dotenv()

SUPABASE_URL = os.getenv("SUPABASE_URL")
SUPABASE_ANON_KEY = os.getenv("SUPABASE_ANON_KEY")
SUPABASE_SERVICE_ROLE_KEY = os.getenv("SUPABASE_SERVICE_ROLE_KEY")


_anon_client = None
_admin_client = None

def get_anon_client() -> Client:
    """Returns a singleton Supabase client with anon key."""
    global _anon_client
    if _anon_client is None:
        _anon_client = create_client(SUPABASE_URL, SUPABASE_ANON_KEY)
    return _anon_client

def get_admin_client() -> Client:
    """Returns a singleton Supabase client with service role key."""
    global _admin_client
    if _admin_client is None:
        _admin_client = create_client(SUPABASE_URL, SUPABASE_SERVICE_ROLE_KEY)
    return _admin_client
