backend/requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
lxml
python-dotenv
supabase
PyJWT
cryptography
playwright
python-multipart
apscheduler
fake-useragent
